API/requirements.txt
Flask-Bcrypt == 1.0.1
Flask-JWT-Extended == 4.4.0
Flask-RESTful == 0.3.9
Flask == 2.1.2
flask-mongoengine == 1.0.0